backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.23.2
pydantic>=2.4.2
geopandas>=0.14.0
numpy>=1.22.0
pandas>=2.0.0
scikit-learn>=1.3.0
python-multipart>=0.0.6
pillow>=10.0.0
rasterio>=1.3.8
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
geopy>=2.4.0
httpx>=0.25.0
matplotlib>=3.7.0
shapely>=2.0.0
flask>=2.3.0 